requirements.txt
tornado
cryptography
elasticsearch
inflect
numpy
pandas
PyYAML
spacy
SQLAlchemy
